Default to IL jurisdiction (#3)
There are [some changes coming in EFSPIntegration](https://github.com/SuffolkLITLab/docassemble-EFSPIntegration/pull/229#issue-1887961043) that require this interview to specifically say that it's in Illinois.

I checked and things still work when you visit the efile account from the Appearance interview, which is how the vast majority of the users will visit it; you'd have to know the exact URL to go to otherwise. But this is a good backup to have.